Add files via upload
The FindResilience code has been modified to now source resilience directly from FishBase. Note that the SciName has to match FishBase exactly now. There is a way to validate names (validate_names), but this just provides a list of names (not the original names), so this is hard to merge. I commented out the parts in this code that aren't needed anymore--using K and tm to determine the resilience, but could be supplemented if desired.